predict.spline now handles the Designmatrix generation for new predictions.
The first part of Simulation 1 is actually done. The ML Estimators for gamma and beta appear to be unbiased.
Strictly speaking gamma appears to be biased since the Estimated Bias is much larger than the SE of the Estimate according to the paper by Morris et al. I suspect the exp link function overblows the Errors, leading to more Error in the bias.